chardev: Let IOReadHandler use size_t type
qemu_chr_be_write_impl uses size_t
all callers use size_t

The number of bytes can not be negative nor zero.

?Qemu has integer overflows because IOReadHandler and its associated functions use a signed integer data type for a size value.

Fixed 2 format string:
- hw/char/spapr_vty.c
- hw/usb/ccid-card-passthru.c

Fixes: CVE-2018-18438
Reported-by: Arash Tohidi Chafi
Suggested-by: Paolo Bonzini <pbonzini@redhat.com>
Acked-by: Alberto Garcia <berto@igalia.com>

+++ hw/rdma/rdma_backend.c
@@ -1020,12 +1020,12 @@ static inline int rdmacm_mux_can_receive(void *opaque)
     return rdmacm_mux_can_process_async(backend_dev);
 }
 
-static void rdmacm_mux_read(void *opaque, const uint8_t *buf, int size)
+static void rdmacm_mux_read(void *opaque, const uint8_t *buf, size_t size)
 {
     RdmaBackendDev *backend_dev = (RdmaBackendDev *)opaque;
     RdmaCmMuxMsg *msg = (RdmaCmMuxMsg *)buf;
 
-    pr_dbg("Got %d bytes\n", size);
+    pr_dbg("Got %zu bytes\n", size);
     pr_dbg("msg_type=%d\n", msg->hdr.msg_type);
     pr_dbg("op_code=%d\n", msg->hdr.op_code);
